+++ opensquirrel/decomposer/cnot_decomposer.py
@@ -4,7 +4,7 @@ import math
 
 from opensquirrel.common import ATOL
 from opensquirrel.decomposer.general_decomposer import Decomposer
-from opensquirrel.decomposer.zyz_decomposer import get_zyz_decomposition_angles
+from opensquirrel.decomposer.zyz_decomposer import ZYZDecomposer
 from opensquirrel.default_gates import CNOT, Ry, Rz, X
 from opensquirrel.ir import BlochSphereRotation, ControlledGate, Float, Gate
 from opensquirrel.merger import general_merger
@@ -39,7 +39,7 @@ class CNOTDecomposer(Decomposer):
 
         # Try special case first, see https://arxiv.org/pdf/quant-ph/9503016.pdf lemma 5.5
         controlled_rotation_times_x = general_merger.compose_bloch_sphere_rotations(X(target_qubit), g.target_gate)
-        theta0_with_x, theta1_with_x, theta2_with_x = get_zyz_decomposition_angles(
+        theta0_with_x, theta1_with_x, theta2_with_x = ZYZDecomposer().get_decomposition_angles(
             controlled_rotation_times_x.angle, controlled_rotation_times_x.axis
         )
         if abs((theta0_with_x - theta2_with_x) % (2 * math.pi)) < ATOL:
@@ -59,7 +59,7 @@ class CNOTDecomposer(Decomposer):
                 + [Rz(q=g.control_qubit, theta=Float(g.target_gate.phase - math.pi / 2))]
             )
 
-        theta0, theta1, theta2 = get_zyz_decomposition_angles(g.target_gate.angle, g.target_gate.axis)
+        theta0, theta1, theta2 = ZYZDecomposer().get_decomposition_angles(g.target_gate.angle, g.target_gate.axis)
 
         A = [Ry(q=target_qubit, theta=Float(theta1 / 2)), Rz(q=target_qubit, theta=Float(theta2))]
